starters/dockerized-django42/tailwind-htmx/{{cookiecutter.project_slug}}/requirements/base-base.txt
#
# This file is autogenerated by pip-compile with Python 3.10
# by the following command:
#
#    pip-compile base-base.in
#
argon2-cffi==23.1.0
    # via -r base-base.in
argon2-cffi-bindings==21.2.0
    # via argon2-cffi
asgiref==3.7.2
    # via django
async-timeout==4.0.2
    # via redis
certifi==2023.5.7
    # via requests
cffi==1.15.1
    # via
    #   argon2-cffi-bindings
    #   cryptography
charset-normalizer==3.2.0
    # via requests
cryptography==41.0.2
    # via pyjwt
defusedxml==0.7.1
    # via python3-openid
dj-database-url==2.1.0
    # via -r base-base.in
django==4.2.5
    # via
    #   -r base-base.in
    #   dj-database-url
    #   django-allauth
django-allauth==0.56.1
    # via -r base-base.in
django-environ==0.11.2
    # via -r base-base.in
idna==3.4
    # via requests
oauthlib==3.2.2
    # via requests-oauthlib
psycopg==3.1.10
    # via -r base-base.in
pycparser==2.21
    # via cffi
pyjwt[crypto]==2.7.0
    # via django-allauth
python-slugify==8.0.1
    # via -r base-base.in
python3-openid==3.2.0
    # via django-allauth
pytz==2023.3
    # via -r base-base.in
redis==5.0.0
    # via -r base-base.in
requests==2.31.0
    # via
    #   django-allauth
    #   requests-oauthlib
requests-oauthlib==1.3.1
    # via django-allauth
sqlparse==0.4.4
    # via django
text-unidecode==1.3
    # via python-slugify
typing-extensions==4.7.1
    # via
    #   asgiref
    #   dj-database-url
    #   psycopg
urllib3==2.0.3
    # via requests
